### 2021-07-04 #3450 @whedon
[PRE REVIEW]: MMM: A C++ library for simulating large herbivores
**Submitting author:** @wtraylor (<a href="http://orcid.org/0000-0002-4813-1072">Wolfgang Traylor</a>)
**Repository:** <a href="https://github.com/wtraylor/modular_megafauna_model" target ="_blank">https://github.com/wtraylor/modular_megafauna_model</a>
**Version:** 1.1.0
