### 2021-07-04 #3450 @Kevin-Mattheus-Moerman
@wtraylor thanks for this submission. I will do an initial scope check and will search for a handling editor for this submission. In the mean time can you work on the following:

- [ ] I suggest you add a plain `LICENSE` file for your repository in the main folder. This way both GitHub and also our systems can detect the open source license for this work.  I see you have a LICENSES folder but perhaps you can keep third party packages only there?
- [ ] Can you check that potentially missing DOI :point_up:
