+++ paper.bib
@@ -43,7 +43,8 @@ https://hds.hebis.de/ubffm/Record/HEB11399558X
   author   = {DeAngelis, Donald L. and Mooij, Wolf M.},
   title    = {In Praise of Mechanistically Rich Models},
   chapter  = {5},
-  pages    = {62--82}
+  pages    = {62--82},
+  doi      = {10.2307/j.ctv1dwq0tq.9}
 }
 
 @article{dirzo2014defaunation,
