### 2021-07-05 6bfc83b
Add license texts and CC0 headers
