+++ paper.bib
@@ -1,3 +1,9 @@
+@Comment{
+SPDX-FileCopyrightText: 2021 Wolfgang Traylor <wolfgang.traylor@senckenberg.de>
+
+SPDX-License-Identifier: CC0-1.0
+}
+
 @article{bokeolen2018estimating,
   title    = {Estimating Grazing Potentials in {Sudan} Using Daily Carbon Allocation in Dynamic Vegetation Model},
   journal  = {Rangeland Ecology \& Management},
