### 2021-07-05 #3450 @wtraylor
@Kevin-Mattheus-Moerman Thank you for receiving my submission.

> Can you check that potentially missing DOI

Done. ✔️

> I suggest you add a plain `LICENSE` file for your repository in the main folder.

My repository follows the [REUSE](https://reuse.software/) standard by the Free Software Foundation Europe. The REUSE convention provides clear license information for each file (compare my [README](https://github.com/wtraylor/modular_megafauna_model/tree/master#license)). My project contains both both source code and documentation, which require different licenses: LGPL for code and CC-BY for docs and graphics (compare [Creative Commons FAQ](https://creativecommons.org/faq/#can-i-apply-a-creative-commons-license-to-software) for software vs. media licensing).

I am afraid that adding a `LICENSE` or `COPYING` file in the root of the repository would create more confusion than clarity because, well, not all of the repository is under *one* license. Yes, unfortunately GitHub does not detect the REUSE standard yet, but my source code that makes up the project is under the LGPL (as indicated by the banner in the `README.md`). Would it be possible to pass this information manually into the JOSS system without the need for a `LICENSE` file?
