### 2021-07-05 #3450 @Kevin-Mattheus-Moerman
@wtraylor thanks for checking that DOI. Also thanks for your comments regarding the REUSE convention. I will have to read up on that. Given that the software caries an OSI approved license this is fully compatible with JOSS so this is fine and I think no changes are needed.
