### 2021-07-07 #3450 @KristinaRiemer
@Kevin-Mattheus-Moerman I wouldn't be able to edit this until mid August, and I don't have any experience with C++, so I don't know if it would be worth trying to find someone else to be the editor?
