### 2021-07-22 bd61b8a
Add CI tests for bad HFT--habitat count combos
