+++ tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_01.toml
@@ -1,0 +1,100 @@
+# SPDX-FileCopyrightText: 2020 W. Traylor <wolfgang.traylor@senckenberg.de>
+#
+# SPDX-License-Identifier: CC0-1.0
+
+# This instruction file has 3 habitats per group, which is not a multiple of 2 HFTs.
+
+[general]
+
+# Number of simulation years. [1,∞)
+years = 100
+
+# Number of habitat groups.
+habitat_groups = 1
+
+# Number of habitats per group.
+habitats_per_group = 3 # ERROR
+
+[environment]
+
+# Monthly mean air temperature [°C].
+# See `grass.daily_decay_rate` for how to use the array.
+air_temperature = [
+  -2, # Jan
+  2,  # Feb
+  7,  # Mar
+  12, # Apr
+  18, # May
+  23, # Jun
+  26, # Jul
+  25, # Aug
+  20, # Sep
+  13, # Oct
+  6,  # Nov
+  0   # Dec
+]
+
+[grass]
+
+# Proportional daily grass decay rate.
+# Each value in the array is for one month. They are cycled. So a one-element
+# array will give a constant decay rate. 12 values will give a repeating annual
+# cycle.
+daily_decay_rate = [ 0.005 ]
+
+# Proportional maximum daily grass growth rate.
+# Example: A value of 0.1 means an increase of 10% every day.
+# See `grass.daily_decay_rate` for how to use the array.
+# 0.03 per day = 0.2 per week (Owen-Smith 2002)
+daily_growth_rate = [
+  0.00, # Jan
+  0.00, # Feb
+  0.00, # Mar
+  0.03, # Apr
+  0.03, # May
+  0.03, # Jun
+  0.03, # Jul
+  0.03, # Aug
+  0.03, # Sep
+  0.00, # Oct
+  0.00, # Nov
+  0.00  # Dec
+]
+
+# Fractional grass digestibility (0,1].
+# See `grass.daily_decay_rate` for how to use the array.
+digestibility = [
+  0.4, # Jan
+  0.4, # Feb
+  0.4, # Mar
+  0.7, # Apr
+  0.7, # May
+  0.7, # Jun
+  0.7, # Jul
+  0.6, # Aug
+  0.5, # Sep
+  0.5, # Oct
+  0.4, # Nov
+  0.4  # Dec
+]
+
+# Fraction of the patch/habitat covered by grass.
+fpc = 0.8
+
+# Initial grass biomass in g/m².
+# Must be smaller than `grass.saturation_mass`.
+initial_mass = 100 # (Owen-Smith 2002)
+
+# Saturation grass biomass in g/m².
+# Must be greater than `grass.initial_mass`.
+saturation_mass = 200 # (Owen-Smith 2002)
+
+# Ungrazeable grass reserve in g/m².
+# Not available to grazers but necessary for regrowth
+ungrazeable_reserve = 20 # (Owen-Smith 2002)
+
+#####################################################################
+# References:
+#
+#	Owen-Smith, N. (2002). A metaphysiological modelling approach to stability in
+#	herbivore–vegetation systems. Ecological Modelling, 149(1), 153-178.
