### 2021-07-23 db491e7
Check value of one_hft_per_habitat for testing
