### 2021-07-23 8e5d74f
Prepare WorldConstructor::create_populations()
Add a counter so we can create one HFT per habitat
