+++ src/Fauna/world_constructor.test.cpp
@@ -26,12 +26,11 @@ TEST_CASE("Fauna::WorldConstructor", "") {
   // prepare HFT list
   HftList hftlist = create_hfts(3, *params);
 
-  WorldConstructor world_cons(params, hftlist);
-
-  SECTION("create_populations() for several HFTs") {
+  SECTION("create_populations() for several HFTs in one habitat") {
     REQUIRE(params->herbivore_type == HerbivoreType::Cohort);
     REQUIRE(params->one_hft_per_habitat == false);
-    PopulationList* pops = world_cons.create_populations();
+    WorldConstructor world_cons(params, hftlist);
+    PopulationList* pops = world_cons.create_populations(0);
     REQUIRE(pops != NULL);
     // Check that there is one population per HFT.
     for (auto& hft : hftlist) {
@@ -47,10 +46,27 @@ TEST_CASE("Fauna::WorldConstructor", "") {
     }
   }
 
+  SECTION("create_populations() for one HFT per habitat") {
+    REQUIRE(params->herbivore_type == HerbivoreType::Cohort);
+    params->one_hft_per_habitat = true;
+    WorldConstructor world_cons(params, hftlist);
+    for (int i = 0; i < 3 * hftlist.size(); i++) {
+      PopulationList* pops = world_cons.create_populations(i);
+      REQUIRE(pops != NULL);
+      REQUIRE(pops->size() == 1);  // Only 1 HFT per habitat!
+      // Check that the *right* HFT is in the populations list.
+      const CohortPopulation& pop = (CohortPopulation&)pops->front();
+      const Hft& found_hft = pop.get_hft();
+      const Hft& right_hft = *hftlist[i % hftlist.size()];
+      REQUIRE(found_hft == right_hft);
+    }
+  }
+
   SECTION("create_populations() for one HFT") {
+    WorldConstructor world_cons(params, hftlist);
     REQUIRE(params->herbivore_type == HerbivoreType::Cohort);
     const auto hft = hftlist[0];
-    PopulationList* pops = world_cons.create_populations();
+    PopulationList* pops = world_cons.create_populations(0);
     REQUIRE(pops != NULL);
     // Check that the CohortPopulation is of the given HFT.
     CHECK(&((CohortPopulation*)pops->begin()->get())->get_hft() == hft.get());
