### 2021-07-23 3161f5b
Add constructor for World for testing w/o insfiles
