### 2021-07-23 cb1e354
Add World::get_sim_units() for testing purpose
