### 2021-07-23 5067fc5
Check that World is activated by default
