### 2021-07-23 aacd872
Implement World::World() constructor for testing
