### 2021-07-23 ea48fab
Check in World::simulate_day() for consecutiveness
I had forgotten to implement that.
