### 2021-07-23 44e099f
Throw exception in get_hfts() if hftlist is NULL
