### 2021-07-23 cf8edd9
Only check output interval after first sim day
This was a bug that the output interval was checked before any output
data had been added. Perhaps I had not discovered the bug because I
always had a spinup time without herbivores
