+++ src/Fauna/world.cpp
@@ -186,7 +186,10 @@ void World::simulate_day(const Date& date, const bool do_herbivores) {
 
   // Write output when it’s ready.
   assert(output_writer.get() != NULL);
-  if (output_aggregator->get_interval().matches_output_interval(
+  assert(last_date.get() != NULL);
+  bool is_first_simulation_day = (date == *last_date);
+  if (!is_first_simulation_day &&
+      output_aggregator->get_interval().matches_output_interval(
           get_params().output_interval))
     for (const auto& datapoint : output_aggregator->retrieve())
       output_writer->write_datapoint(datapoint);
