### 2021-07-23 c1d55b1
Increment date outside of CHECK_NOTHROW()
