### 2021-07-23 4df3f69
Allow calls to simulate_day() without sim. units
My previous bugfix (cf8edd92858fd5275c9228635f327e13fa0a2bb0) didn’t
address the issue.
