### 2021-07-23 415291c
Set World::last_date on every call in simulate_day()
