### 2021-07-23 14270b3
Revert "Increment date outside of CHECK_NOTHROW()"
This reverts commit c1d55b18e6c47338ef5160d55cb179dadb3f901d.
