### 2021-07-23 dbee455
Fix bad pointers in Fauna::World unit test
For that I made create_hfts() return a pointer. That makes it easier to
put the result into a std::shared_ptr object.
