+++ tests/dummy_hft.h
@@ -16,22 +16,20 @@
 
 namespace Fauna {
 /// Create a simple, valid HftList
-inline HftList create_hfts(const int count, const Parameters& params) {
-  HftList hftlist;
+inline HftList* create_hfts(const int count, const Parameters& params) {
+  HftList* hftlist = new HftList;
   for (int i = 0; i < count; i++) {
     std::shared_ptr<Hft> hft(new Hft());
     // construct name for HFT
-    std::ostringstream stream;
-    stream << "hft" << i;
-    hft->name = stream.str();
+    hft->name = "hft" + std::to_string(i);
 
     std::string msg;
     if (!hft->is_valid(params, msg))
       FAIL("create_hfts(): HFT is not valid:\n" << msg);
 
-    hftlist.push_back(hft);
+    hftlist->push_back(hft);
   }
-  REQUIRE(hftlist.size() == count);
+  REQUIRE(hftlist->size() == count);
   return hftlist;
 }
 }  // namespace Fauna
