### 2021-07-23 fadd86e
Implement counter for habitats for creating herbis
