### 2021-07-23 a1bb1b7
Remove superfluous comments
