+++ .gitlab-ci.yml
@@ -228,15 +228,15 @@ demo_toml_files:
   script:
     - (! ./build/megafauna_demo_simulator
       "tests/good_insfiles/two_hfts_01.toml"
-      "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts-01.toml");
+      "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_01.toml");
     - "rm -f *.tsv"
     - (! ./build/megafauna_demo_simulator
       "good_insfiles/two_hfts_01.toml"
-      "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts-02.toml");
+      "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_02.toml");
     - "rm -f *.tsv"
     - (! ./build/megafauna_demo_simulator
       "tests/good_insfiles/two_hfts_01.toml"
-      "tests/good_demosim_insfiles/habitats_is_multiple_of_hfts-01.toml");
+      "tests/good_demosim_insfiles/habitats_is_multiple_of_hfts_01.toml");
 
 demo_simulation:
   image: gcc:latest
