### 2021-07-23 94487ec
Reduce simulation time of test demo simulations
This will shorten CI a bit
