+++ tests/good_demosim_insfiles/habitats_is_multiple_of_hfts_01.toml
@@ -7,7 +7,7 @@
 [general]
 
 # Number of simulation years. [1,∞)
-years = 100
+years = 10
 
 # Number of habitat groups.
 habitat_groups = 2
