+++ tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_02.toml
@@ -8,7 +8,7 @@
 [general]
 
 # Number of simulation years. [1,∞)
-years = 100
+years = 10
 
 # Number of habitat groups.
 habitat_groups = 2
