### 2021-07-23 3b9f888
Correct exception type for wrong argument
