### 2021-07-26 f4b1952
Add unit test for habitat count as multiple of HFT
