### 2021-07-26 50b1f8d
Add unit test for same habitat count per agg unit
