### 2021-07-26 64627a6
Remove negation of return value for *good* check
