+++ .gitlab-ci.yml
@@ -234,9 +234,9 @@ demo_toml_files:
       "good_insfiles/two_hfts_01.toml"
       "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_02.toml");
     - "rm -f *.tsv"
-    - (! ./build/megafauna_demo_simulator
+    - ./build/megafauna_demo_simulator
       "tests/good_insfiles/two_hfts_01.toml"
-      "tests/good_demosim_insfiles/habitats_is_multiple_of_hfts_01.toml");
+      "tests/good_demosim_insfiles/habitats_is_multiple_of_hfts_01.toml";
 
 demo_simulation:
   image: gcc:latest
