### 2021-07-26 de9a7ce
Implement counting habitats per aggregation unit
