### 2021-07-26 d366886
Fix linter: create output writer not for linting
