### 2021-07-27 31998df
Plot only existing files in demo_results.Rmd
close gh-39
