+++ tools/demo_simulator/demo_results.Rmd
@@ -6,9 +6,15 @@ SPDX-License-Identifier: CC0-1.0
 
 # Modular Megafauna Output
 
-```{r include = FALSE}
+```{r setup, include = FALSE}
 library(ggplot2)
 library(tidyr)
+
+# "timestep" is the output interval: a day, a year, (roughly) a month, or a decade.
+get_timestep <- function(df){
+  unique_years.v <- sort(unique(df$year))
+  timestep <- abs(unique_years.v[2] - unique_years.v[1])
+}
 
 # READ TABLES
 
@@ -43,33 +49,15 @@ read_per_hft_per_forage_table <- function(filename){
   gather(tbl, key = HFT, value = VALUE, !all_of(fixed_vars))
 }
 
-# Read individual tables
-
-available.tbl <- read_per_forage_type_table("available_forage.tsv")
-available.tbl$VALUE <- available.tbl$VALUE * 10^-6 # convert kg/km² to kg/m²
-
-bodyfat.tbl <- read_per_hft_table("body_fat.tsv")
-
-eaten.tbl <- read_per_hft_per_forage_table("eaten_forage_per_ind.tsv")
-
-eaten_nitrogen.tbl <- read_per_hft_table("eaten_nitrogen_per_ind.tsv")
-
-mass.tbl <- read_per_hft_table("mass_density.tsv")
-mass.tbl$VALUE <- mass.tbl$VALUE / 100 # convert kg/km² to kg/ha
-
-inddens.tbl <- read_per_hft_table("individual_density.tsv")
-inddens.tbl$VALUE <- inddens.tbl$VALUE / 100 # convert ind/km² to ind/ha
-
-# `timestep` is the output interval: a day, a year, (roughly) a month, or a decade.
-unique_years.v <- sort(unique(mass.tbl$year))
-timestep <- abs(unique_years.v[2] - unique_years.v[1])
-
 # The first of each month as Julian day: for drawing month axis labels.
 julian_month_days <- julian(as.Date(paste(1:12, "1 1970"), "%m %d %Y"))
 ```
 
-```{r echo = FALSE}
+```{r mass_density, echo = FALSE, eval = file.exists("mass_density.tsv")}
+mass.tbl <- read_per_hft_table("mass_density.tsv")
+mass.tbl$VALUE <- mass.tbl$VALUE / 100 # convert kg/km² to kg/ha
 # We plot the points not in the beginning of each time step, but in the middle.
+timestep <- get_timestep(mass.tbl)
 ggplot(mass.tbl, aes(x = year + timestep / 2, y = VALUE, color = HFT)) +
   stat_summary(fun = mean, geom = "line") +
   coord_cartesian(xlim = c(min(mass.tbl$year), max(mass.tbl$year) + timestep)) +
@@ -90,8 +78,11 @@ ggplot(mass.tbl, aes(x = year + timestep / 2, y = VALUE, color = HFT)) +
   )
 ```
 
-```{r echo = FALSE}
+```{r individual_density, echo = FALSE, eval = file.exists("individual_density.tsv")}
+inddens.tbl <- read_per_hft_table("individual_density.tsv")
+inddens.tbl$VALUE <- inddens.tbl$VALUE / 100 # convert ind/km² to ind/ha
 # We plot the points not in the beginning of each time step, but in the middle.
+timestep <- get_timestep(inddens.tbl)
 ggplot(inddens.tbl, aes(x = year + timestep / 2, y = VALUE, color = HFT)) +
   stat_summary(fun = mean, geom = "line") +
   coord_cartesian(
[…]
