### 2021-07-27 30565b3
Allow one_hft_per_habitat only for "Cohorts"
