+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog][] by Olivier Lacan, and this project a
 ### Added
 - `demo_results.Rmd` now only tries to plot those files that are available [#39]
 
+### Fixed
+- Allow `simulation.one_hft_per_habitat` only for `herbivore_type = "Cohort"`
+
 ## [1.1.1] - 2021-07-26
 
 ### Fixed
