### 2021-07-27 aa1ec3c
Check if HFTs were correctly created in habitats
