### 2021-07-27 0905fd2
Add unit test for one_hft_per_habitat == false
