### 2021-07-27 045843c
Remove redundant check for herbivore_type = cohort
