### 2021-07-27 5d1d44e
Fix non-functional one_hft_per_habitat
The counter for how many habitats were already created in an aggregation
unit was not correctly incremented because I compared char* with each
other.
