### 2021-07-27 6f69336
Tidy up Fauna::World
Remove World::is_activated() and dummy constructor World::World()
because they are not used at all.
