### 2021-07-27 7331322
Fix output if one_hft_per_habitat==true
Fauna::Output::CombinedData::merge() had a bad for loop to check the
HFTs/groups from the to-be-merged other object. This way the data from
the `other` object always got dropped.
