+++ CHANGELOG.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog][] by Olivier Lacan, and this project a
 
 ### Fixed
 - Allow `simulation.one_hft_per_habitat` only for `herbivore_type = "Cohort"`
-- Issue that only one HFT got created if `simulation.one_hft_per_habitat = true`
+- Issue that only one HFT got created and outputted if `simulation.one_hft_per_habitat = true`
 
 ## [1.1.1] - 2021-07-26
 
