### 2021-07-28 7427a15
Restore World::World() and World::is_activated()
In commit 6f69336 (version 1.1.2) I had changed the library interface.
That was a bad idea. LPJ-GUESS didn’t work anymore. Since MMM is
supposed to be backward-compatible, I restore the old library interface,
but mark it as deprecated.
