### 2021-07-28 6e0d292
Declare reset day option in World::simulate_day()
