+++ src/Fauna/world.test.cpp
@@ -28,16 +28,16 @@ TEST_CASE("FAUNA::World", "") {
   REQUIRE_NOTHROW(World(PARAMS, HFTLIST));
   CHECK_THROWS(World(PARAMS, HFTLIST).create_simulation_unit(NULL));
   CHECK_NOTHROW(World(PARAMS, HFTLIST).get_params());
-  CHECK_NOTHROW(World(PARAMS, HFTLIST).simulate_day(Date(1, 2), true));
-  CHECK_NOTHROW(World(PARAMS, HFTLIST).simulate_day(Date(1, 2), false));
+  CHECK_NOTHROW(World(PARAMS, HFTLIST).simulate_day(Date(1, 2)));
+  CHECK_NOTHROW(World(PARAMS, HFTLIST).simulate_day(Date(1, 2)));
 
   SECTION("Dummy Constructor (deprecated)") {
     World w;
     REQUIRE(!w.is_activated());
     CHECK_THROWS(w.get_params());
     CHECK_THROWS(w.create_simulation_unit(NULL));
-    CHECK_NOTHROW(w.simulate_day(Date(1, 2), true));
-    CHECK_NOTHROW(w.simulate_day(Date(1, 2), false));
+    CHECK_NOTHROW(w.simulate_day(Date(1, 2)));
+    CHECK_NOTHROW(w.simulate_day(Date(1, 2)));
   }
 
   SECTION("Unit test constructor") {
@@ -50,28 +50,54 @@ TEST_CASE("FAUNA::World", "") {
 
   SECTION("simulate_day()") {
     World world(PARAMS, HFTLIST);
-    world.simulate_day(Date(0, 0), true);
+    world.simulate_day(Date(0, 0));
     // Run for some days
-    CHECK_NOTHROW(world.simulate_day(Date(1, 0), true));
-    CHECK_NOTHROW(world.simulate_day(Date(2, 0), true));
-    CHECK_NOTHROW(world.simulate_day(Date(3, 0), true));
+    CHECK_NOTHROW(world.simulate_day(Date(1, 0)));
+    CHECK_NOTHROW(world.simulate_day(Date(2, 0)));
+    CHECK_NOTHROW(world.simulate_day(Date(3, 0)));
     // Try to simulate the same day again.
-    CHECK_THROWS(world.simulate_day(Date(3, 0), true));
+    CHECK_THROWS(world.simulate_day(Date(3, 0)));
     // Try to simulate the previous day again
-    CHECK_THROWS(world.simulate_day(Date(2, 0), true));
+    CHECK_THROWS(world.simulate_day(Date(2, 0)));
     // Try to simulate some bad arbitrary days.
-    CHECK_THROWS(world.simulate_day(Date(4, 1), true));
-    CHECK_THROWS(world.simulate_day(Date(14, 0), true));
-    CHECK_THROWS(world.simulate_day(Date(14, 3), true));
-    CHECK_THROWS(world.simulate_day(Date(10, 3), true));
-    CHECK_THROWS(world.simulate_day(Date(10, 3), true));
+    CHECK_THROWS(world.simulate_day(Date(4, 1)));
+    CHECK_THROWS(world.simulate_day(Date(14, 0)));
+    CHECK_THROWS(world.simulate_day(Date(14, 3)));
+    CHECK_THROWS(world.simulate_day(Date(10, 3)));
+    CHECK_THROWS(world.simulate_day(Date(10, 3)));
 
     // Check that the year boundary works.
     for (int day = 4; day < 365; day++)
-      CHECK_NOTHROW(world.simulate_day(Date(day, 0), true));
-    CHECK_NOTHROW(world.simulate_day(Date(0, 1), true));
[…]
