### 2021-07-28 9145471
Implement World::SimDayOptions::reset_date
