### 2021-08-02 8c4a4ba
Add test TOML file without any HFTs
