### 2021-08-02 c44bf3e
Add CI tests for one_hft_per_habitat without HFTs
