+++ .gitlab-ci.yml
@@ -230,6 +230,11 @@ demo_toml_files:
       "tests/good_insfiles/two_hfts_01.toml"
       "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_01.toml");
     - "rm -f *.tsv"
+      # Without HFTs, the option "one_hft_per_habitat" has no effect.
+    - ./build/megafauna_demo_simulator
+      "tests/good_insfiles/no_hfts_02.toml"
+      "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_01.toml");
+    - "rm -f *.tsv"
     - (! ./build/megafauna_demo_simulator
       "good_insfiles/two_hfts_01.toml"
       "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_02.toml");
@@ -238,8 +243,9 @@ demo_toml_files:
       "tests/good_insfiles/two_hfts_01.toml"
       "tests/good_demosim_insfiles/habitats_is_multiple_of_hfts_01.toml";
     - "rm -f *.tsv"
+      # A simulation should run without HFTs and still produce output.
     - ./build/megafauna_demo_simulator
-      "tests/good_insfiles/no_hfts.toml"
+      "tests/good_insfiles/no_hfts_01.toml"
       "tests/good_demosim_insfiles/minimal.toml";
 
 demo_simulation:
