### 2021-08-02 81ac1ce
Add unit tests for Fauna::World without HFTs
