### 2021-08-03 e3275fa
Fix typo in doxygen reference
