### 2021-08-03 2d7f04d
Fix bad doxygen reference
