### 2021-08-03 68f1b4e
Fix syntax error in CI
