+++ .gitlab-ci.yml
@@ -233,7 +233,7 @@ demo_toml_files:
       # Without HFTs, the option "one_hft_per_habitat" has no effect.
     - ./build/megafauna_demo_simulator
       "tests/good_insfiles/no_hfts_02.toml"
-      "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_01.toml");
+      "tests/bad_demosim_insfiles/habitats_not_multiple_of_hfts_01.toml";
     - "rm -f *.tsv"
     - (! ./build/megafauna_demo_simulator
       "good_insfiles/two_hfts_01.toml"
