### 2021-08-03 6b8c2f5
Add CI test to not create HFT output without HFTs
