### 2021-08-04 #3450 @dirmeier
Ciao @diehlpk , the scope of the paper is pretty much outside of my own research area. I could review the C++ code,  but regarding the impact, novelty, etc. I would need to rely on the other reviewers, if that is ok with you.
Cheers, Simon
