### 2021-08-04 35e2090
Create HFT output table files only if there’re HfTs
