### 2021-08-11 #3450 @diehlpk
> > @wtraylor Can you please recommend some potential reviewers, especially with C++ experience?
>
> Outside my own research group, I know that [Phillip Papastefanou](https://www.lsai.wzw.tum.de/en/team/phillip-papastefanou/) from the Technical University Munich is competent in C++ and mechanistic models. He’s a late-stage PhD student and works with LPJ-GUESS, too, the same vegetation model that I couple MMM to. Otherwise we are not personally or professionally affiliated with each other.
>
> If PhD student level is considered not experienced enough, perhaps there is someone else in [Anja Rammig](https://www.lsai.wzw.tum.de/en/team/prof-dr-anja-rammig/)’s working group available; they’re working with complex models. I think [Katrin Fleischer](https://www.lsai.wzw.tum.de/en/team/dr-katrin-fleischer/) might have some C++ skills.

@wtraylor I contacted more JOSS reviewers, we have many for the topic, however, not so many with C++ experience. Can you recommend more reviewers? Unfortunately, none of the above responded.
