### 2021-08-12 #3450 @CELavecchia
Hi,

Sorry but I can’t do it.

On Wed, 11 Aug 2021 at 23:45, Nicholas Nadeau, Ph.D., P.Eng. <
***@***.***> wrote:

> Hey @CELavecchia <https://github.com/CELavecchia> @martinmodrak
> <https://github.com/martinmodrak> @nnadeau <https://github.com/nnadeau>
> @fernandoandreotti <https://github.com/fernandoandreotti> @expectopatronum
> <https://github.com/expectopatronum> @betteridiot
> <https://github.com/betteridiot> @Rgcjk would you be interested in
> reviewing this paper?
>
> Unfortunately I don't have time at the moment.
>
