### 2021-08-13 #3450 @wtraylor
nnadeau wrote:
> @wtraylor I contacted more JOSS reviewers, we have many for the topic, however, not so many with C++ experience. Can you recommend more reviewers? Unfortunately, none of the above responded.

I am tapping my network and will post some suggestions next week.
