### 2021-08-16 #3450 @wtraylor
I can suggest some folks at Lund University, Sweden, who are involved in the development of the dynamic global vegetation model [LPJ-GUESS](https://web.nateko.lu.se/lpj-guess/index.html), which MMM integrates with. They work with mechanistic ecological models, though on the vegetation side. All of them have experience in C++ programming.
- [Paul Miller](https://portal.research.lu.se/portal/en/persons/paul-miller(f0ccbb8b-12e7-4e35-9dec-799053c3c695).html)
- [Stefan Olin](https://portal.research.lu.se/portal/en/persons/stefan-olin(7d8f5a19-3600-4922-a135-0ea43d3ee1c2).html)
- [David Wårlind](https://portal.research.lu.se/portal/en/persons/david-waarlind(bc722af4-cb49-4e16-9f61-85c82c860ab4).html)
- [Lars Nieradzik](https://portal.research.lu.se/portal/en/persons/lars-nieradzik(9405fe73-d27a-4c21-ab56-32427260ee82).html)
