### 2021-08-16 #3450 @martinmodrak
Hi,
I am pretty much out of my depth regarding the subject matter, but can potentially review the software engineering aspects. However, definitely not on a tight schedule, I have some upcoming deadlines and most likely won't be able to devote time for the review before mid September.
