### 2021-08-17 #3450 @wtraylor
A friend of mine recommended two potential reviewers who bring both ecological background and C++ skills:
- [Gavin Simpson](http://www.iecs-uregina.ca/people/index.html) – <https://orcid.org/0000-0002-9084-8413>
- [Ethan White](https://www.weecology.org/people/ethan-white/) – <https://orcid.org/0000-0001-6728-7745>

Both are busy professors with probably very limited time for reviews, unfortunately …
