### 2021-08-17 3352a52
Fix low digestibility values in output
Digestibility was merged for output in `ForageBase::merge` with wrong
weights, which didn’t consider zero forage values. Digestibility only
applies to actual forage, so zero forage means zero weight.
