### 2021-08-18 7b7e988
Set naming scheme for prereleases
