### 2021-08-18 #3631 @wtraylor
> Failed to discover a valid open source license.

See https://github.com/openjournals/joss-reviews/issues/3450#issuecomment-874160752
