### 2021-09-02 #3631 @diehlpk
> I currently don't have access to a computer where I can compile and run the code, not sure when I'm allowed back in the office, but it won't be this month (September). Den ons 1 sep. 2021 kl 17:28 skrev whedon ***@***.***>:
> […](#)
> 👋 @StefanOlin, please update us on how your review is going (this is an automated reminder). — You are receiving this because you were mentioned. Reply to this email directly, view it on GitHub, or unsubscribe. Triage notifications on the go with GitHub Mobile for iOS or Android.
> -- Stefan Olin Researcher Department of Physical Geography and Ecosystem Science Lund University ***@***.*** ***@***.***

@StefanOlin Thanks for the update.
