### 2021-09-12 wtraylor/modular_megafauna_model#51 @dirmeier
MMM paper review
Ciao Wolfgang,

I'll use this issue for the paper review. The paper is well written, but as you might have read, I don't have much insight in your field of research, so I only have contentual comments.

Regarding the paper, I made some minor suggestions which you are free to adopt or not. Other than that, I think it would be nice to add the mathematical formulation of the mechanistic model. What do you think about that?

I uploaded the review [here](https://www.dropbox.com/s/rhnk0fpu92pnxv0/10.21105.joss.03631-r0.pdf?dl=0).

Cheers,
Simon
