### 2021-09-12 wtraylor/modular_megafauna_model#52 @dirmeier
MMM code review
Ciao Wolfgang,

the code/documentation/unit tests/etc etc are excellent and you've put significant work into it, so every review on this would feel a bit pedantic to me.

I'd like to ask you, however, to
-  add the example building/usage into README.md such that one doesn't have to go through the html docs,
- add CMAKE as dependency,
- add "add C++11 compatible compiler" as dependency.

Cheers, S
