### 2021-10-04 #3631 @dirmeier
Ciao @diehlpk, you can find my reviews here: https://github.com/wtraylor/modular_megafauna_model/issues/51, https://github.com/wtraylor/modular_megafauna_model/issues/52

Cheers, S
