### 2021-10-29 5c3a0f9
Move quickstart guide from Doxygen to README.md
Suggested by Simon Dirmeier in gh-52.
