+++ README.md
@@ -16,6 +16,8 @@ Modular Megafauna Model
 
 [REUSE]: https://reuse.software
 
+
+
 Overview
 --------
 
@@ -48,18 +50,106 @@ This project follows the [Pitchfork Layout](https://github.com/vector-of-bool/pi
 - `tests/`: Unit tests and test scripts.
 - `tools/`: Different helper tools for the developer.
 
-Usage
------
+Quickstart
+----------
 
-The first step in working with the model is to create the documentation and read the introductory pages.
-On the Quickstart page you will find instructions for compiling the source code.
+### Compile and Run Unit Tests
 
-### Compile the Doxygen Documentation
+To check if the modular megafauna model works by itself correctly, you should compile it and run the unit tests.
+You will need [Cmake](https://cmake.org) (3.10 or later) and a C++ compiler supporting the C++11 standard.
+It should work well with GCC, the [GNU C++ compiler](https://gcc.gnu.org) (4.8 or later).
+
+Open a Unix shell in the root of this repository and run:
+
+```sh
+mkdir -p build
+cd build
+cmake -DBUILD_TESTING=ON ..
+make megafauna_unit_tests
+./megafauna_unit_tests
+```
+
+If the unit tests all pass successfully, you are good to go!
+
+### Run the Demo Simulator
+
+With the CMake option `BUILD_DEMO_SIMULATOR=ON` you can compile the demo simulator.
+This independent program is a very simple grass simulator that hosts the megafauna library.
+It should work out of the box with the instructions files in the `examples/` folder.
+
+Again, open a terminal in the repository root folder and run:
+
+```sh
+mkdir -p build
+cd build
+cmake -DBUILD_DEMO_SIMULATOR=ON ..
+make megafauna_demo_simulator
+./megafauna_demo_simulator "../examples/megafauna.toml" "../examples/demo_simulation.toml"
+```
+
+Congratulations, you have run your first simulation!
+
+The output files are tab-separated text (`.tsv`) files.
+The simulator does not want to have you lose your results and will refuse to overwrite existing output files.
+So if you want to run it again, you will first need to remove the previously created output tables: `rm *.tsv`
+
+A very simple [RMarkdown](https://rmarkdown.rstudio.com/) file to visualize the demo output is provided in the `build` folder.
+It is called `demo_results.Rmd`.
+You can open and render it in [RStudio](https://www.rstudio.com/).
+
+Alternatively, you can compile the RMarkdown file into an HTML file without RStudio.
+You will need to have [R](https://www.r-project.org/) (version 3, 4, or later) installed.
+
+Open a terminal in the `build` directory and execute `R` to get into an interactive R session.
+In the R console, you execute the following commands:
+
+```r
+install.packages("rmarkdown")
+library(rmarkdown)
+render("demo_results.md")
+quit(save = "no")
+```
+
+This should have produced the HTML file `demo_results.html`, which you can open in your web browser.
+The results should show a rising population curve, which means that the example herbivore has been able to survive and reproduce.
+
+Feel free to play with the MMM parameters in `megafauna.toml` and the Demo Simulator parameters in `demo_simulation.toml`.
+
+### Customize the Instruction File
+
+The Modular Megafauna Model library requires a single instruction file to set general simulation parameters and herbivore (HFT = Herbivore Functional Type) parameters.
+The instruction file is in [TOML v0.5][] format; see there for a description of the syntax.
+String parameters are generally case insensitive.
+
+[TOML v0.5]: https://github.com/toml-lang/toml/tree/v0.5.0
+
+All possible options are listed in the example file under [examples/megafauna.toml](examples/megafauna.toml).
+You will find all parameters explained in detail in the Doxygen documentation of the two classes [Fauna::Parameters][] and [Fauna::Hft][].
[…]
