### 2021-10-29 ce21b35
Auto-install required R packages
This makes it easy for first-time users who have just installed R.
