### 2021-10-29 8d7b885
Fix typo in code in quickstart guide
