### 2021-10-29 3e5d4d4
Add Pandoc as dependency
