+++ .gitlab-ci.yml
@@ -281,13 +281,14 @@ demo_simulation:
     expire_in: 20 minutes
 
 plot_demo_results:
+  # This job should match the quickstart guide in the README.md.
   image: r-base:latest
   stage: postprocess
   needs: ["demo_simulation"]
   before_script:
     - apt-get update -qq
     - apt-get install -y -qq pandoc
-    - Rscript -e 'install.packages(c("ggplot2", "rmarkdown", "tidyr"))'
+    - Rscript -e 'install.packages("rmarkdown")'
   script:
     - cd build
     - Rscript -e "rmarkdown::render('demo_results.Rmd')"
