### 2021-11-11 c90a690
Amend colloquial formulation + add other use-cases
