### 2021-11-16 wtraylor/modular_megafauna_model#52 @wtraylor
@dirmeier Thank you for the suggestions. I have moved the Quickstart Guide to the README and have mentioned the dependencies in the README, too. You can see it in the [develop branch](https://github.com/wtraylor/modular_megafauna_model/tree/develop).

Let me know what you think.
