### 2021-11-16 1cf7f08
Make "none" plural
Suggested in gh-51
