### 2021-11-16 wtraylor/modular_megafauna_model#51 @wtraylor
> Regarding the paper, I made some minor suggestions which you are free to adopt or not.

Thanks. I have adopted them.

In a PDF comment you wrote:

>  MMM simulates potential probability mass distributions of population counts of mammoths…

Actually, MMM does not output probability distributions but discrete population counts. It’s (mostly) a deterministic model. Uncertainty or probability distributions can be calculated in a sensitivity analysis by varying input parameters. I am working on such an analysis for my herbivores-of-interest, but it’s a separate set of scripts.

---

> Other than that, I think it would be nice to add the mathematical formulation of the mechanistic model. What do you think about that?

There are two obstacles to adding the mathematical formulation (model description) in the JOSS paper:

1. The modularity of MMM means that there is no *one* model that could be formulated. It’s a bit like Lego. You put together the model as you need it. That’s what I see as MMM’s core strength, which is why I have only included an [Ecological Model Discussion](https://modular-megafauna-model.readthedocs.io/en/latest/page_model.html) in the documentation, rather than writing an actual model description.
1. A full model description would by far exceed the scope of a JOSS paper.

For the actual application of my model for my dissertation I have written a complete model description (in [ODD format](http://doi.org/10.18564/jasss.4259)). When that is published, others can use it as a template according to how they have set up or extended MMM for their research question.
