### 2021-11-16 #3631 @wtraylor
> How is the editing going?

@diehlpk Thanks for checking in. I have adopted the (minor) changes suggested by @dirmeier.

> Could you might recommend some additional reviewers?

I think I have tapped most of my network already, but I will try to come up with other suggestions.
