### 2021-11-30 #3631 @dilawar
MMM is a very high-quality library with an excellent user and API documentation. The repository follows state-of-the-art DevOps. The library is a step towards a critical and hard problem of validating assumptions and discovering missing parameters while modeling a complex system. The library can also serve as a very useful pedagogical tool in Ecology class (I'd have certainly enjoyed seeing some simulations of important ideas).

The author has designed the library to be (C++) user-friendly by choosing a modern build system (CMake). I also appreciate in-source dependencies and choosing an easy-to-read TOML over often used but hardly human-readable XML for user-program interaction.

The CI/CD pipeline is comprehensive. It not only tests for the correctness, but also for memory leaks and lint errors. Tooling is provided for code-formatting.

Contribution guidelines are clear. The author has also set up a social channel (matrix) to engage with the users.

It is heartening to an academic software built with so much care. I wholeheartedly recommend publication of this library.

Following are a few more comments and a question.

- _Using first time:_ It was a breeze to build the library. All CMake options (i.e., `-DBUILD_DOC` and `-DBUILD_TESTING`) worked as expected. Demo also ran fine. After the run, I was asked to install R to see the plots. I installed R and required dependencies (including Pandoc) and managed to see the plots.  Perhaps, the author would like to add expected results (plots) to the demo documentation.

- [ ] Following lines from the paper has confused me:

  > Here, MMM is coupled with LPJ-GUESS [refs], using daily grass growth (Boke-Olén et al., 2018). LPJ-GUESS is proprietary software and not publicly available.

  Is LPJ-GUESS is a hard dependency? I guess not because I could run `megafauna_demo_simulator` without installing LPJ-GUESS. However, I am under the impression that it is an optional dependency and can be used along with HMM? If yes,  then there is no indication of it in the documentation. Moreover, the educational version (binary) of LPJ-GUESS is only available for Windows. It is not clear how LPJ-GUESS and MMM will interact with each (if they do).
