### 2021-12-02 0b616d5
Rewrite confusing paragraph about LPJ-GUESS
Compare
https://github.com/openjournals/joss-reviews/issues/3631#issuecomment-982513731
