+++ paper.md
@@ -59,10 +59,14 @@ The currently implemented model concepts originate in large parts from @pachzelt
 Later studies have implemented conceptually similar grazer models for other DGVMs: @dangal2017integrating for DLEM and @zhu2018large for ORCHIDEE.
 However, to my knowledge, none of these implementations are reusable across different vegetation models.
 
-The author is currently using MMM to simulate potential densities of mammoths, steppe bison, and horse in the last ice age.
-Here, MMM is coupled with LPJ-GUESS [@smith2001representation; @smith2014implications], using daily grass growth [@bokeolen2018estimating].
-LPJ-GUESS is proprietary software and not publicly available.
-MMM could also be parametrized for extant herbivores, such as Scandinavian reindeer or Mongolian gazelles.
+The author is currently using MMM integrated with LPJ-GUESS [@smith2001representation; @smith2014implications], using daily grass growth [@bokeolen2018estimating].[^lpj_guess]
+Here, LPJ-GUESS simulates grass as forage for herbivores.
+This particular project is concerned with simulating potential densities of mammoths, steppe bison, and horse in the last ice age.
+However, MMM could also be parametrized for extant herbivores, such as cattle, Scandinavian reindeer, or Mongolian gazelles.
+MMM can also be coupled with dynamic vegetation models other than LPJ-GUESS.
+
+[^lpj_guess]:
+The proprietary LPJ-GUESS source code, including the MMM-coupled version, may be available upon request at Lund University (<https://iis4.nateko.lu.se/lpj-guess/>).
 
 # Features
 
