### 2021-12-02 f32347b
Clarify that LPJ-GUESS educational is uncompatible
Compare:
https://github.com/openjournals/joss-reviews/issues/3631#issuecomment-982513731
