### 2021-12-02 14a2d29
Refer to CI artifacts for expected demo output
Compare:
https://github.com/openjournals/joss-reviews/issues/3631#issuecomment-982513731
