### 2021-12-02 #3631 @wtraylor
@dilawar Thank you very much for your review and remarks. I have tried to rephrase the paragraph about LPJ-GUESS coupling in a clearer way. In the README, I mentioned that the educational version of LPJ-GUESS is not compatible with MMM.

> Perhaps, the author would like to add expected results (plots) to the demo documentation.

I also thought about that. However, the example parameters and the plot script might change. After editing them, one could forget to update the screenshot of the expected demo output in the documentation. Then it can easily happen that the user’s output is different than the one in the docs, and the user might try to find an error where there is none. I didn’t come up with a good way to automate that, so I added a reference to the CI output.
