### 2021-12-02 #3631 @diehlpk
@wtraylor I found some time and read your paper. I can recommend it for acceptance if you add this last change.

- [ ] `Through the instruction file, users can... ` I think it might be helpful to put a link (https://github.com/wtraylor/modular_megafauna_model/tree/master/examples) to the examples here. Or maybe add to the end of this paragraph: Example input files can be found here or whatever you prefer.
