### 2021-12-06 a999e25
Add footnote linking to example instruction files
Compare
https://github.com/openjournals/joss-reviews/issues/3631#issuecomment-984847254
