### 2021-12-06 #3631 @wtraylor
@diehlpk I have added a footnote pointing to the example instruction files in the code.

I am currently debugging the CI for the documentation so I can create a release afterwards. I guess it makes sense to have the version that’s published in JOSS identifiable with a release tag.
