### 2021-12-06 #3631 @diehlpk
@wtraylor Yes, please make a new version which has to include all the changes, you made during the JOSS review.

Please do the following

- [ ] Post the new version here.
- [ ] Upload this version to Zenodo or FigShare and post the corresponding DOI here. Note that the title and the authors have to match with the PDF.
