### 2021-12-14 d0322af
Add linke to Doxygen manual in comment
