+++ docs/CMakeLists.txt
@@ -20,6 +20,9 @@ if (DOXYGEN_FOUND)
     set (DOXYGEN_WARN_AS_ERROR NO)
   endif()
 
+  # The following CMake options correspond to Doxygen configuration file
+  # options, which are documented here:
+  # https://www.doxygen.nl/manual/config.html
   set (DOXYGEN_CITE_BIB_FILES "bibliography.bib")
   set (DOXYGEN_DISTRIBUTE_GROUP_DOC YES)
   set (DOXYGEN_EXAMPLE_PATH "../examples")
