### 2021-12-14 4a59c01
Disable `dot` in generating Doxygen documentation
Trying to generate the Doxygen documentation in the
readthedocs/build:latest container, I had received various error
messages like this:

```
sh: 1: /usr/sbin/dot: not found
/home/wtraylor/modular_megafauna_model/clone/mmm/build/docs/html/classFauna_1_1invalid__option__coll__graph.dot:1: error: Problems running dot: exit code=127, command='/usr/s
bin/dot', arguments='"/home/wtraylor/modular_megafauna_model/clone/mmm/build/docs/html/classFauna_1_1invalid__option__coll__graph.dot" -Tpng -o "/home/wtraylor/modular_megafa
una_model/clone/mmm/build/
[…]
