+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog][] by Olivier Lacan, and this project a
 - Exception if started without HFTs
 - Now HFT table files won’t be created if there are no HFTs defined.
 - Too low digestibility values in output, due to wrong weights in aggregation.
+- CI doxygen generation failing with error "sh: 1: /usr/sbin/dot: not found". `dot` is now disabled.
 
 ## [1.1.2] - 2021-07-27
 
