+++ CMakeLists.txt
@@ -8,7 +8,7 @@ cmake_minimum_required (VERSION 3.10)
 project ("Modular Megafauna Model"
   # This version must match exactly the Git tag!
   # Compare the "check_version" job in ".gitlab-ci.yml".
-  VERSION 1.1.3
+  VERSION 0.0.0
   DESCRIPTION "A physiological, dynamic herbivore simulator in C++."
   LANGUAGES CXX
   )
