### 2021-12-14 #3631 @wtraylor
@diehlpk I was able to fix the CI and released [**version 1.1.3**](https://github.com/wtraylor/modular_megafauna_model/releases/tag/1.1.3) containing all changes from this review. It’s archived on Zenodo: [10.5281/zenodo.5779267](https://doi.org/10.5281/zenodo.5779267)
