### 2021-12-14 #3631 @diehlpk
@wtraylor The paper title

`MMM: A C++ library for simulating large herbivores`

does not match the title of the Zenodo archive

`Modular Megafauna Model`

Please change that, so I can accept the paper.
